refactor catalogs

+++ rust/frontend/src/catalog/mod.rs
@@ -7,8 +7,8 @@ pub(crate) mod database_catalog;
 mod schema_catalog;
 pub(crate) mod table_catalog;
 
-pub(crate) type DatabaseId = u64;
-pub(crate) type SchemaId = u64;
+pub(crate) type DatabaseId = u32;
+pub(crate) type SchemaId = u32;
 pub(crate) type TableId = risingwave_common::catalog::TableId;
 pub(crate) type ColumnId = risingwave_common::catalog::ColumnId;
